# api-spec.yaml
openapi: 3.0.3
info:
  title: TestCase Management API
  version: 1.0.0
  description: |
    TestCase 管理系統 API
    
    - 主要前端流程以 Product/Version 為入口，再進入 Module 與 TestCase
    - 使用 SQLite 資料庫
    - 支援 JSON 格式資料交換
    - 提供完整的 CRUD 操作
servers:
  - url: http://127.0.0.1:5000
    description: 本地開發伺服器
...
paths:
...
  /api/admin/reset:
    post:
      tags:
        - System
      summary: 重置所有資料與 ID 序列
      description: |
        清空所有資料表（products, modules, test_cases, testruns, testrun_test_cases）
        並重置 ID 序列計數器。
        
        ⚠️ **警告**：此操作不可逆轉，所有資料將永久刪除！
        
        用途：
        - 從 GitHub 下載後重新開始
        - 開發環境重置
        - 測試環境清理
      responses:
        '200':
          description: 重置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== Products ====================
  /api/products:
    get:
      tags:
        - Products
      summary: 取得所有 Product/Version 列表
      operationId: listProducts
      responses:
        '200':
          description: 成功取得產品列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Products
      summary: 建立新產品
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: 產品建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/products/{id}:
    get:
      tags:
        - Products
      summary: 取得特定產品
      operationId: getProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: 成功取得產品
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Products
      summary: 更新產品
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: 產品更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Products
      summary: 刪除產品
      operationId: deleteProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: 產品刪除成功
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Modules ====================
  /api/products/{product_id}/modules:
    get:
      tags:
        - Modules
      summary: 取得指定 Product/Version 底下的所有模組
      operationId: listModules
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: 成功取得模組列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Module'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - Modules
      summary: 建立新模組
      operationId: createModule
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleCreate'
      responses:
        '201':
          description: 模組建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Module'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/modules/{id}:
    get:
      tags:
        - Modules
      summary: 取得特定模組
      operationId: getModule
      parameters:
        - $ref: '#/components/parameters/ModuleId'
      responses:
        '200':
          description: 成功取得模組
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Module'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Modules
      summary: 更新模組
      operationId: updateModule
      parameters:
        - $ref: '#/components/parameters/ModuleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleUpdate'
      responses:
        '200':
          description: 模組更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Module'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Modules
      summary: 刪除模組
      operationId: deleteModule
      parameters:
        - $ref: '#/components/parameters/ModuleId'
      responses:
        '200':
          description: 模組刪除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== TestCases ====================
  /api/testcases:
    get:
      tags:
        - TestCases
      summary: 取得所有 TestCase（平面列表）
      operationId: listTestCases
      description: 未提供分頁參數時回傳完整列表；提供 page 或 per_page 時由資料庫使用 LIMIT/OFFSET 分頁。
      parameters:
        - name: q
          in: query
          description: Case 名稱模糊查詢
          schema:
            type: string
        - name: page
          in: query
          description: 頁碼，必須為正整數
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: 每頁筆數
          schema:
            type: integer
            enum: [20, 50, 100]
            default: 20
      responses:
        '200':
          description: 成功取得 TestCase 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - TestCases
      summary: 建立新 TestCase
      operationId: createTestCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseCreate'
      responses:
        '201':
          description: TestCase 建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/testcases/{id}:
    get:
      tags:
        - TestCases
      summary: 取得特定 TestCase
      operationId: getTestCase
      parameters:
        - $ref: '#/components/parameters/TestCaseId'
      responses:
        '200':
          description: 成功取得 TestCase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - TestCases
      summary: 更新 TestCase
      operationId: updateTestCase
      parameters:
        - $ref: '#/components/parameters/TestCaseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestCaseUpdate'
      responses:
        '200':
          description: TestCase 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - TestCases
      summary: 刪除 TestCase
      operationId: deleteTestCase
      parameters:
        - $ref: '#/components/parameters/TestCaseId'
      responses:
        '200':
          description: TestCase 刪除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/testcases/hierarchy:
    get:
      tags:
        - TestCases
      summary: 取得 Product/Version > Module > TestCase 階層資料
      operationId: listTestCaseHierarchy
      responses:
        '200':
          description: 成功取得 TestCase 階層資料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestCaseHierarchyResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== TestRuns ====================
  /api/testruns:
    get:
      tags:
        - TestRuns
      summary: 取得所有 TestRun
      operationId: listTestRuns
      responses:
        '200':
          description: 成功取得專案列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TestRun'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - TestRuns
      summary: 建立新 TestRun
      operationId: createTestRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestRunCreate'
      responses:
        '201':
          description: 專案建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRun'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      description: 產品 ID
      schema:
        type: integer
        example: 1
...
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "產品A"
        created_at:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00"
    ProductCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 128
          example: "產品A"
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 128
          example: "產品A（新名稱）"

    # Module
...
    ModuleCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 128
          example: "登入模組"
...
    ResetResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
          example: "success"
        message:
          type: string
          example: "已重置所有資料與 ID 序列"
        warning:
          type: string
          example: "所有資料已清空，ID 序列已重置，系統已重新初始化"
  responses:
    BadRequest:
      description: 請求資料無效
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Invalid request data"
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found"
    InternalError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error"